Streamlit/Enter1/requirements.txt
streamlit
pandas
supabase
requests
plotly
openai
fpdf2
